moved jobs to lib
