+++ docker-compose.yml
@@ -41,7 +41,7 @@ services:
       - ./umich_catalog_indexing/.env
       - ./umich_catalog_indexing/.env-dev-values
       - ./mock-ht/.env-dev-values
-    command: bundle exec sidekiq -r ./bin/jobs.rb
+    command: bundle exec sidekiq -r ./lib/jobs.rb
 
   redis:
     image: 'redis:6.2'
