+++ umich_catalog_indexing/lib/jobs.rb
